Update x86/fbranch sample to fix issues with NaNs.
To reproduce the new binary, compile fbranch.c with 'clang -ffast-math -m32'.

+++ tests/recompilation-tests/fbranch.c
@@ -16,7 +16,7 @@ int main()
     float b;
 
     scanf("%f", &b);
-    printf("a is %f, b is %f\n", a, b);
+    printf("a is %f, b is %f\n", (double)a, (double)b);
 
     if (a == b) printf("Equal\n");
     if (a != b) printf("Not Equal\n");
